# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go QRCode Generator CMS - Swagger API Discovery",
        "title": "Go QRCode Generator CMS - Swagger API Discovery",
        "contact": {
            "name": "Xuan Hoa Le",
            "email": "dso.intern.xuanhoa@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/qrcodes": {
            "get": {
                "description": "Find QRCode by custom condition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qrcodes"
                ],
                "summary": "Find QRCode by custom condition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Version",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Error Level",
                        "name": "errorLevel",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Start Time",
                        "name": "startTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "End Time",
                        "name": "endTime",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.standardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.standardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.standardResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create QRCode using custom configuration",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qrcodes"
                ],
                "summary": "Create QRCode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "background",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "borderWidth",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "name": "circleShape",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "content[]",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maximum": 4,
                        "minimum": 1,
                        "type": "integer",
                        "name": "errorLevel",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "foreground",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "name": "transparentBackground",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "-",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Logo",
                        "name": "logo",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Halftone",
                        "name": "halftone",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.standardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.standardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.standardResponse"
                        }
                    }
                }
            }
        },
        "/qrcodes/{qrCodeUUID}": {
            "get": {
                "description": "Find QRCode using UUID - full QRCode Information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qrcodes"
                ],
                "summary": "Find QRCode by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "QRCode UUID",
                        "name": "qrCodeUUID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.standardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.standardResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entity.standardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.standardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.standardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "statusText": {
                    "type": "string"
                }
            }
        },
...
    }
}